MILESTONE : Added Skill Scaffold
- Copied and adapted the views from ProgrammingSkill scaffold
- Change :ranking to :level (less cryptic)
-> Had to modify the star_tag reliying on the :ranking attribute of the
item if was called with; instead add a .star_tag method to both Skill
and ProgrammingSkill models
- Completely refactored the skill_type attribute in Skills :
-- Instead of being an integer, it is now a string that is used to order
the skills. The form allows to either use an old (read: already used)
skill types, or to define a new one.
=> Both very powerful and flexible
- Skills have a slightly different rendering then other Listable content, so we need to fork to
  two different partial here (listable_index_skills and
listable_index_others)
=> Works like a charm

Ccl: All in all, the refactoring seems to work and is ready to be
debugged on real case scenarii.